+++ lib/rubocop/cop/style/negated_while.rb
@@ -30,7 +30,9 @@ module RuboCop
         def autocorrect(node)
           @corrections << lambda do |corrector|
             condition, _body, _rest = *node
-            # unwrap the negated portion of the condition (a send node)
+            # Look inside parentheses around the condition, if any.
+            condition, _ = *condition while condition.type == :begin
+            # Unwrap the negated portion of the condition (a send node).
             pos_condition, _method, = *condition
             corrector.replace(
               node.loc.keyword,